修正部分在win7 32bitｈ的問題

+++ set-IEasDefault.ps1
@@ -7,19 +7,23 @@
 param($runadmin)
 
 function Set-IEasDefault {
+    $check_win10 = (Get-WmiObject -Class Win32_OperatingSystem | Select-Object -Property Version).Version -like '10*'
 
-    Write-Output "在Edge中以IE開啟網站設為永不"
-    Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Edge\IEToEdge" -Name "RedirectionMode" -Value 0 -Force
+    if ($check_win10) {
+        Write-Output "在Edge中以IE開啟網站設為永不"
+        Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Edge\IEToEdge" -Name "RedirectionMode" -Value 0 -Force
 
-    Write-Output "關閉IE中的IEtoEdga元件"
-    Set-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Policies\Ext\CLSID" -Name "{1FD49718-1D00-4B19-AF5F-070AF6D5D54C}" -Value 0 -Force
+        Write-Output "關閉IE中的IEtoEdga元件"
+        Set-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Policies\Ext\CLSID" -Name "{1FD49718-1D00-4B19-AF5F-070AF6D5D54C}" -Value 0 -Force
 
-    #預設IE瀏覽器無法在本機群組原則套用. Registry也無法指定IE, 只能先清除.
-    Write-Output "清除預設瀏覽器設定,讓使用者第一次執行自行選擇"
-    Remove-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Windows\Shell\Associations\UrlAssociations\http\UserChoice" -Name "Progid" -ErrorAction SilentlyContinue
-    Remove-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Windows\Shell\Associations\UrlAssociations\https\UserChoice" -Name "Progid" -ErrorAction SilentlyContinue
+        #預設IE瀏覽器無法在本機群組原則套用. Registry也無法指定IE, 只能先清除.
+        Write-Output "清除預設瀏覽器設定,讓使用者第一次執行自行選擇"
+        Remove-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Windows\Shell\Associations\UrlAssociations\http\UserChoice" -Name "Progid" -ErrorAction SilentlyContinue
+        Remove-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Windows\Shell\Associations\UrlAssociations\https\UserChoice" -Name "Progid" -ErrorAction SilentlyContinue
+    }
 
 }
+
